complete mobile app

+++ site/src/commonMain/kotlin/org/example/blogmultiplatform/models/Post.kt
@@ -8,7 +8,7 @@ data class Post(
     @SerialName("_id")
     val id: String = "",
     val author: String = "",
-    val date: Long = 0L,
+    val date: Double = 0.0,
     val title: String,
     val subtitle: String,
     val thumbnail: String,
@@ -24,7 +24,7 @@ data class PostWithoutDetails(
     @SerialName("_id")
     val id: String = "",
     val author: String,
-    val date: Long,
+    val date: Double,
     val title: String,
     val subtitle: String,
     val thumbnail: String,
